% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cw_download}
\alias{cw_download}
\title{Download from Climate Watch Data}
\usage{
cw_download(data_sources = c("CAIT", "PIK", "UNFCCC"),
  gases = c("All GHG", "CO2", "CH4", "N2O", "F-Gas"), sectors = c(),
  regions, start_year, end_year, sort_col, sort_dir = c("ASC", "DESC"))
}
\arguments{
\item{sort_dir}{}
}
\description{
Download from Climate Watch Data
}
